+++ index.html
@@ -10,8 +10,8 @@
 
         gtag('config', 'UA-140160470-1');
     </script>
-    <script src="dist/interact.js" type="module"></script>
-    <script src="dist/enums.js" type="module"></script>
+    <script src="dist/src/interact.js" type="module"></script>
+    <script src="dist/src/enums.js" type="module"></script>
     <link rel="stylesheet" type="text/css" href="styles.css">
     <title>Destroyer</title>
 </head>
