+++ tsconfig.json
@@ -25,6 +25,6 @@
     "**/*.tsx",
     "**/*.js",
     "**/*.jsx"
-  ],
+, "jest.config.mjs"  ],
   "exclude": ["node_modules", "dist"]
 }
